perf(masterFeeHandler): pack Config slot, dedupe setShares loops
- Pack TokenFeeConfigLib.Config slot 0 (isSplit + uint16 totalDirectBps),
  making totalDirectBps a warm SLOAD in _depositSplit / _depositSingle.
- Simplify _depositSingle: drop dead isDirect flag, branch on totalDirectBps.
- Move setShares' claimable-snapshot loop into _setSharesInternal's wipe
  loop (single pass over claimableRecipients, ethPerBps cached); drop the
  redundant _claimedPerBps write on old recipients.
- _populateNewShares: cache ethPerBps; skip the SSTORE 0->0 on
  _claimedPerBps at registration.

+++ src/libraries/TokenFeeConfigLib.sol
@@ -5,8 +5,9 @@ library TokenFeeConfigLib {
     uint256 internal constant BPS_TOTAL = 10_000;
 
     struct Config {
+        // Packed into slot 0: bool (1 byte) + uint16 (2 bytes), 29 bytes free.
         bool isSplit;
-        uint256 totalDirectBps;
+        uint16 totalDirectBps; // bounded by BPS_TOTAL = 10_000, fits in uint16.
         uint256 ethPerBps;
         address[] directReceivers;
         address[] claimableRecipients;
